refactor: extract Wigner corep theory into separate module, fix UNI numbering
- Add src/irrep/wigner.rs: Wigner test, little-group filter, character table
  builder with full LaTeX documentation
- Add identify_unitary_subgroup(uni) → SG number of unitary subgroup H
- Add uni_from_bns/uni_from_og for BNS/OG label → UNI lookup
- Add compute_coreps(bns, k_label) high-level API: BNS+k-label → all coreps
- Fix: compute_corepresentation now uses H's irreps (not parent SG's)
- Fix: only unitary magnetic ops need H mapping; anti-unitary ops use None
- Fix UNI 1073→1066: 1073 is the Litvin number, 1066 is the actual UNI
- Verified: 128.406 (UNI 1066) unitary subgroup = SG 118 against BCS
- Change u16→usize for UNI, SG numbers, domain/arm counts
- Document ISOTROPY mag_nlabel index = UNI number in CLAUDE.md

+++ src/irrep/types.rs
@@ -227,7 +227,7 @@ impl IsotropyRecord {
 #[derive(Debug, Clone, Copy)]
 pub struct IsotropyRecord {
     /// Subgroup space group number (1–230)
-    pub sg: u16,
+    pub sg: usize,
     /// Hermann-Mauguin symbol
     pub symbol: &'static str,
     /// Schoenflies symbol
@@ -235,9 +235,9 @@ pub struct IsotropyRecord {
     /// Order-parameter direction label
     pub direction: &'static str,
     /// Number of domains
-    pub domains: u16,
+    pub domains: usize,
     /// Number of arms in the star
-    pub arms: u16,
+    pub arms: usize,
 }
 
 /// A magnetic isotropy subgroup: the lower-symmetry magnetic space group
@@ -246,8 +246,8 @@ pub struct IsotropyRecord {
 /// the possible magnetic structures that can form.
 #[derive(Debug, Clone, Copy)]
 pub struct MagneticIsotropyRecord {
-    /// Magnetic space group number (1–1651, ISOTROPY Suite numbering)
-    pub mag_sg: u16,
+    /// Magnetic space group UNI number (1–1651)
+    pub mag_sg: usize,
     /// BNS (Belov-Neronova-Smirnova) symbol, e.g. `"Pm'mm"`
     pub bns_label: &'static str,
     /// ISOTROPY label, e.g. `"47.251"`
